make: change Enet build to only run as needed
Define the Enet target as an actual file to ensure that make knows when
it already exists (and does not need to be rebuilt).

Add an include of the enet/Makefile.am to ensure that make knows the
file dependencies of Enet, and can ensure rebuild upon modification of
the Enet source files.


Former-commit-id: d44014f0b9def873b536e4219b3641c15cd4689a
Former-commit-id: 46d8e438673f958ce9c42de27095d1853f0143fd